Merge: back- dev

See merge request s12-final/S12P31B106!266

+++ backend/src/main/java/com/voiceprint/backend/api/auth/dto/DiaryResponse.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 //@AllArgsConstructor
 @Slf4j
 public class DiaryResponse {
-    private Long diaryId;
+    private Integer diaryId;
     private String title;
     private String content;
     private LocalDateTime createdAt;
