+++ backend/src/main/java/com/voiceprint/backend/domain/Repository/ChatbotRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChatbotRepository extends JpaRepository<Chatbot, Long> {
+public interface ChatbotRepository extends JpaRepository<Chatbot, Byte> {
 
 }
