+++ backend/src/main/java/com/voiceprint/backend/domain/Repository/ProfileImageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfileImageRepository extends JpaRepository<ProfileImage, Long> {
+public interface ProfileImageRepository extends JpaRepository<ProfileImage, Byte> {
 }
